tests/wpt/web-platform-tests/content-security-policy/style-src/style-src-3_2.html

<!DOCTYPE HTML>
<html>
<head>
    <title>Inline style attributes should not be applied without 'unsafe-inline' style-src directive.</title>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <script>
      onload = function() {
        test(function() {
        var text = document.getElementById("content");
        assert_true(getComputedStyle(text).marginLeft != "2px", "Inline style attribute should not be applied to text");
       });
      }
    </script>
</head>
<body>
    <h1>Inline style attributes should not be applied without 'unsafe-inline' style-src directive.</h1>
    <div id='log'></div>

    <div id="content" style="margin-left: 2px">This text should not have a margin-left of 2</div>

    <script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=style-src%20%27self%27'></script>

</body>
</html>
